--- src/by/suboch/Cake.java
@@ -1,5 +1,0 @@
-package by.suboch;
-
-public class Cake {
-
-}
