--- src/by/suboch/Chair.java
@@ -1,8 +1,0 @@
-package by.suboch;
-
-public class Chair extends StandingFurniture {
-
-    void stand(){
-        System.out.println("Я стул, я стою");
-    }
-}
