--- src/by/suboch/Cup.java
@@ -1,7 +1,0 @@
-package by.suboch;
-
-public class Cup extends FlyingDishes implements FlyingObjects {
-    public void flies() {
-        System.out.println("Я чашка, я взлетела");
-    }
-}
