--- src/by/suboch/Cupboard.java
@@ -1,8 +1,0 @@
-package by.suboch;
-
-public class Cupboard extends StandingFurniture {
-
-    void stand() {
-        System.out.println("Я шкаф, я стою");
-    }
-}
