--- src/by/suboch/Fork.java
@@ -1,7 +1,0 @@
-package by.suboch;
-
-public class Fork extends FlyingDishes implements FlyingObjects {
-    public void flies() {
-        System.out.println("Я вилка, я взлетела");
-    }
-}
