--- src/by/suboch/Info.java
@@ -1,7 +1,0 @@
-package by.suboch;
-
-public interface Info {
-
-    void showInfo();
-    void printName();
-}
