--- src/by/suboch/Mixer.java
@@ -1,8 +1,0 @@
-package by.suboch;
-
-public class Mixer extends StandingFurniture {
-
-    void stand() {
-        System.out.println("Я смеситель, я смешиваю");
-    }
-}
