--- src/by/suboch/Pasta.java
@@ -1,4 +1,0 @@
-package by.suboch;
-
-public class Pasta {
-}
