--- src/by/suboch/Plate.java
@@ -1,8 +1,0 @@
-package by.suboch;
-
-public class Plate extends FlyingDishes implements FlyingObjects {
-
-    public void flies() {
-        System.out.println("Я тарелка, я взлетела");
-    }
-}
