--- src/by/suboch/Skinali.java
@@ -1,7 +1,0 @@
-package by.suboch;
-
-public class Skinali extends StandingFurniture {
-    void stand() {
-
-    }
-}
