--- src/by/suboch/Stand.java
@@ -1,8 +1,0 @@
-package by.suboch;
-
-public abstract class Stand {
-
-    static void standing() {
-        System.out.println("Мы все сожем стоять");
-    }
-}
