--- src/by/suboch/TVset.java
@@ -1,23 +1,0 @@
-package by.suboch;
-
-import suboch.interfaces.Info;
-
-public class TVset implements Info {
-    public String name;
-
-    public TVset(String name) {
-        this.name = name;
-    }
-
-
-    @Override
-    public void showInfo() {
-
-    }
-
-    @Override
-    public void printName() {
-
-    }
-
-}
