--- src/by/suboch/Table.java
@@ -1,13 +1,0 @@
-package by.suboch;
-
-public class Table extends StandingFurniture {
-
-    public static void standing() {
-        System.out.println("Я стол, я стою");
-    }
-
-    @Override
-    void stand() {
-
-    }
-}
