--- src/suboch/Chair.java
@@ -1,10 +1,0 @@
-package suboch;
-
-import by.suboch.Stand;
-
-public class Chair extends Stand {
-
-    static void standing(){
-        System.out.println("Я стул, я стою");
-    }
-}
