--- src/suboch/Cupboard.java
@@ -1,10 +1,0 @@
-package suboch;
-
-import by.suboch.Stand;
-
-public class Cupboard extends Stand {
-
-    static void standing() {
-        System.out.println("Я шкаф, я стою");
-    }
-}
