--- src/suboch/Fly.java
@@ -1,11 +1,0 @@
-package suboch;
-
-import suboch.interfaces.Flying;
-
-public abstract class  Fly implements Flying {
-
-    public void flying() {
-
-    }
-
-}
