--- src/suboch/Mixer.java
@@ -1,10 +1,0 @@
-package suboch;
-
-import by.suboch.Stand;
-
-public class Mixer extends Stand {
-
-    static void standing() {
-        System.out.println("Я смеситель, я смешиваю");
-    }
-}
