--- src/suboch/Skinali.java
@@ -1,9 +1,0 @@
-package suboch;
-
-import by.suboch.Stand;
-
-public class Skinali extends Stand {
-    static void standing() {
-
-    }
-}
