--- src/suboch/Table.java
@@ -1,12 +1,0 @@
-package suboch;
-
-import by.suboch.Stand;
-
-public class Table extends Stand {
-
-
-    static void standing() {
-        System.out.println("Я стол, я стою");
-
-    }
-}
