+++ src/suboch/Test.java
@@ -1,7 +1,9 @@
 package suboch;
 
-import by.suboch.*;
-import suboch.interfaces.Flying;
+import suboch.furniture.Mixer;
+import suboch.interfaces.Info;
+import suboch.technics.Lamp;
+import suboch.technics.MicrowaveOven;
 
 public class Test {
     public static void main(String[] args) {
@@ -9,17 +11,12 @@ public class Test {
         Lamp lamp1 = new Lamp(100);
         MicrowaveOven MO = new MicrowaveOven("LG");
         Info lamp2 = (Info) new Lamp(200);
-        Info TVset = (Info) new TVset("Horizont");
-        Stand table = new Table();
-        Skinali skinali = new Skinali();
-        Flying cup1 = (Flying) new Cup();
         System.out.println("Изменение в develop");
 
-        cup1.flying();
+
         lamp1.watt();
         lamp1.showInfo();
         lamp2.showInfo();
-        TVset.showInfo();
         MO.showInfo();
         Mixer.standing();
