+++ src/suboch/dishes/Cup.java
@@ -1,10 +1,10 @@
 package suboch.dishes;
 
-import suboch.interfaces.Flying;
+import suboch.dishes.FlyingDishes;
+import suboch.interfaces.FlyingObjects;
 
-public class Cup extends Fly implements Flying {
-
-    public void flying() {
+public class Cup extends FlyingDishes implements FlyingObjects {
+    public void flies() {
         System.out.println("Я чашка, я взлетела");
     }
 }
