--- src/suboch/dishes/Fly.java
@@ -1,8 +1,0 @@
-package suboch.dishes;
-
-import suboch.interfaces.Flying;
-
-public abstract class Fly implements Flying {
-
-    public abstract void flying();
-    }
