+++ src/suboch/dishes/Plate.java
@@ -1,10 +1,11 @@
 package suboch.dishes;
 
-import suboch.interfaces.Flying;
+import suboch.dishes.FlyingDishes;
+import suboch.interfaces.FlyingObjects;
 
-public class Plate extends Fly implements Flying {
+public class Plate extends FlyingDishes implements FlyingObjects {
 
-    public void flying() {
+    public void flies() {
         System.out.println("Я тарелка, я взлетела");
     }
 }
