+++ src/suboch/furniture/Chair.java
@@ -2,7 +2,7 @@ package suboch.furniture;
 
 public class Chair extends StandingObject {
 
-    public void standing() {
+    public static void standing() {
         System.out.println("Я стул, я стою");
     }
 }
