+++ src/suboch/furniture/Cupboard.java
@@ -2,7 +2,7 @@ package suboch.furniture;
 
 public class Cupboard extends StandingObject {
 
-    public void standing() {
+    public static void standing() {
         System.out.println("Я шкаф, я стою");
     }
 }
