+++ src/suboch/furniture/Mixer.java
@@ -2,7 +2,7 @@ package suboch.furniture;
 
 public class Mixer extends StandingObject {
 
-    public void standing() {
+    public static void standing() {
         System.out.println("Я смеситель, я смешиваю");
     }
 }
