+++ src/suboch/furniture/Skinali.java
@@ -1,7 +1,7 @@
 package suboch.furniture;
 
 public class Skinali extends StandingObject {
-    public void standing() {
+    public static void standing() {
 
     }
 }
