+++ src/suboch/furniture/StandingFurniture.java
@@ -1,4 +1,4 @@
-package suboch;
+package suboch.furniture;
 
 public abstract class StandingFurniture {
     abstract void stand();
