+++ src/suboch/furniture/StandingObject.java
@@ -1,5 +1,7 @@
 package suboch.furniture;
 
 public abstract class StandingObject {
-    public abstract void standing();
+    public static void standing() {
+    }
+
 }
