--- src/suboch/interfaces/Flying.java
@@ -1,6 +1,0 @@
-package suboch.interfaces;
-
-public interface Flying {
-
-    void flying();
-}
