+++ src/suboch/interfaces/Info.java
@@ -1,7 +1,7 @@
 package suboch.interfaces;
 
 public interface Info {
-    abstract void showInfo();
 
-    void name();
+    void showInfo();
+    void printName();
 }
