+++ src/suboch/technics/TVset.java
@@ -16,8 +16,8 @@ public class TVset implements Info {
     }
 
     @Override
-    public void name() {
-        System.out.println("Name is " + this.name);
+    public void printName() {
 
     }
+
 }
